Return the 'namelist' data source as a tibble

Returns the included data source namelist as a tibble, by default filtered according to English names and shortnames.

read_namelistmon(
  path = pkgdatasource_path("textdata/namelist", ".yml"),
  file = "namelist",
  lang = "en"
)

Arguments

path
Location of the data source. The default is to use the location of the data source as delivered by the installed package.

file
The filename of the data source, without extension. The default is to use the file delivered by the installed package.

lang
An IETF BCP 47 language tag, such as "en" or "nl", to specify the language of name and shortname to be returned in the tibble. If lang = "all", the full namelist tibble is returned, i.e. containing all languages.

Value

The namelist data frame as a tibble, filtered according to the lang argument. See namelist for documentation of the tibble's contents.

Details

namelist is a data source in the vc-format which provides names and (optionally) shortnames for IDs/codes used in other data sources.

read_namelistmon() reads it and returns it as a tibble. A tibble is a data frame that makes working in the tidyverse a little easier. By default, the data version delivered with the package is used and only English names (lang = "en") are returned.

Recommended usage

read_namelistmon()

read_namelistmon(lang = "nl")

See also

Examples

read_namelistmon()
#> # A tibble: 108 × 4
#>    code            lang  name                                          shortname
#>    <chr>           <chr> <chr>                                         <chr>    
#>  1 aq              en    aquatic                                       NA       
#>  2 ATM             en    Atmospheric monitoring                        NA       
#>  3 ATM_03.1        en    Atmospheric monitoring: environmental pressu… ATM: 03.…
#>  4 ATM_03.1_group1 en    very sensitive non-forest types               very sen…
#>  5 ATM_03.1_group2 en    sensitive non-forest types                    sensitiv…
#>  6 ATM_03.1_group3 en    less to not sensitive non-forest types        less/not…
#>  7 ATM_03.1_group4 en    very sensitive forest types                   very sen…
#>  8 ATM_03.1_group5 en    sensitive forest types                        sensitiv…
#>  9 ATM_03.1_group6 en    less to not sensitive forest types            less/not…
#> 10 ATM_04.1        en    Atmospheric monitoring: environmental pressu… ATM: 04.…
#> # ℹ 98 more rows
read_namelistmon(lang = "nl")
#> # A tibble: 108 × 4
#>    code            lang  name                                          shortname
#>    <chr>           <chr> <chr>                                         <chr>    
#>  1 aq              nl    aquatisch                                     NA       
#>  2 ATM             nl    Atmosferisch meetnet                          NA       
#>  3 ATM_03.1        nl    Atmosferisch meetnet: milieudruk 03.1_eutr_a… ATM: 03.…
#>  4 ATM_03.1_group1 nl    zeer gevoelige niet-bostypes                  zeer gev…
#>  5 ATM_03.1_group2 nl    gevoelige niet-bostypes                       gevoelig…
#>  6 ATM_03.1_group3 nl    minder tot niet gevoelige niet-bostypes       minder/n…
#>  7 ATM_03.1_group4 nl    zeer gevoelige bostypes                       zeer gev…
#>  8 ATM_03.1_group5 nl    gevoelige bostypes                            gevoelig…
#>  9 ATM_03.1_group6 nl    minder tot niet gevoelige bostypes            minder/n…
#> 10 ATM_04.1        nl    Atmosferisch meetnet: milieudruk 04.1_verzu_… ATM: 04.…
#> # ℹ 98 more rows